third_party/WebKit/LayoutTests/media/video-currentTime-set-with-controller.html

<!doctype html>
<html>
    <head>
        <title>Setting currentTime with a media controller present</title>
    </head>
    <body>
        <video mediagroup="x"></video>
        <script src="video-test.js"></script>
        <script>
            testExpected("video.controller instanceof MediaController", true);
            testDOMException("video.currentTime = 0", "DOMException.INVALID_STATE_ERR");
            endTest();
        </script>
    </body>
</html>
